+++ l10n/ga.js
@@ -127,5 +127,6 @@ OC.L10N.register("linkboard", {
     "Maximum number of services reached (%s)": "Maximum number of services reached (%s)",
     "Display": "Display",
     "Show scrollbar": "Show scrollbar",
-    "avg {ms} ms · max {max} ms · {time}": "avg {ms} ms · max {max} ms · {time}"
+    "avg {ms} ms · max {max} ms · {time}": "avg {ms} ms · max {max} ms · {time}",
+    "Invalid format": "Invalid format"
 }, "nplurals=5; plural=(n==1 ? 0 : n==2 ? 1 : n<7 ? 2 : n<11 ? 3 : 4);");