+++ l10n/ja.js
@@ -127,5 +127,6 @@ OC.L10N.register("linkboard", {
     "Maximum number of services reached (%s)": "サービスの最大数に達しました（%s）",
     "Display": "Display",
     "Show scrollbar": "Show scrollbar",
-    "avg {ms} ms · max {max} ms · {time}": "avg {ms} ms · max {max} ms · {time}"
+    "avg {ms} ms · max {max} ms · {time}": "avg {ms} ms · max {max} ms · {time}",
+    "Invalid format": "Invalid format"
 }, "nplurals=1; plural=0;");